refactor: Renamed TextRenderers as ClientRenderer since their purpose serve as rendering for the console client
